Move getConfigurationHash to CrawlerConfigurationUtility
