+++ Tests/Unit/Controller/CrawlerControllerTest.php
@@ -24,8 +24,11 @@ use AOE\Crawler\Service\PageService;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\Test;
+use PHPUnit\Framework\MockObject\Exception;
 use PHPUnit\Framework\MockObject\MockObject;
 use Psr\Log\NullLogger;
+use ReflectionException;
+use ReflectionMethod;
 use TYPO3\TestingFramework\Core\Unit\UnitTestCase;
 
 /**
@@ -131,33 +134,6 @@ class CrawlerControllerTest extends UnitTestCase
         self::assertEquals($expected, $crawlerController->getUrlsForPageRow($pageRow, $skipMessage));
     }
 
-    #[Test]
-    public function getConfigurationHash(): void
-    {
-        $crawlerController = $this->getAccessibleMock(CrawlerController::class, [], [], '', false);
-
-        $configuration = [
-            'paramExpanded' => 'extendedParameter',
-            'URLs' => 'URLs',
-            'NotImportantParameter' => 'value not important',
-        ];
-
-        $originalCheckSum = md5(serialize($configuration));
-
-        self::assertNotNull(
-            $crawlerController->_call('getConfigurationHash', $configuration)
-        );
-
-        self::assertNotEquals(
-            $originalCheckSum,
-            $crawlerController->_call('getConfigurationHash', $configuration)
-        );
-
-        unset($configuration['paramExpanded'], $configuration['URLs']);
-        $newCheckSum = md5(serialize($configuration));
-        self::assertEquals($newCheckSum, $crawlerController->_call('getConfigurationHash', $configuration));
-    }
-
     public static function getUrlsForPageRowDataProvider(): iterable
     {
         yield 'Message equals false, returns Urls from getUrlsForPages()' => [
@@ -187,60 +163,6 @@ class CrawlerControllerTest extends UnitTestCase
             'pageRow' => ['uid' => 'string'],
             '$skipMessage' => 'PageUid "string" was not an integer',
             'expected' => [],
-        ];
-    }
-
-
-    #[DataProvider('getConfigurationHasReturnsExpectedValueDataProvider')]
-    #[Test]
-    public function getConfigurationHasReturnsExpectedValue(array $configuration, string $expected): void
-    {
-        $crawlerLib = $this->getAccessibleMock(CrawlerController::class, [], [], '', false);
-
-        self::assertEquals($expected, $crawlerLib->_call('getConfigurationHash', $configuration));
-    }
-
-    public static function getConfigurationHasReturnsExpectedValueDataProvider(): iterable
-    {
-        yield 'Configuration with either paramExpanded nor URLs set' => [
-            'configuration' => [
-                'testKey' => 'testValue',
-                'paramExpanded' => '',
-                'URLs' => '',
-            ],
-            'expected' => 'a73d2e7035f7fa032237c8cf0eb5be22',
-        ];
-        yield 'Configuration with only paramExpanded set' => [
-            'configuration' => [
-                'testKey' => 'testValue',
-                'paramExpanded' => 'Value not important',
-                'URLs' => '',
-            ],
-            'expected' => 'a73d2e7035f7fa032237c8cf0eb5be22',
-        ];
-        yield 'Configuration with only URLS set' => [
-            'configuration' => [
-                'testKey' => 'testValue',
-                'paramExpanded' => '',
-                'URLs' => 'Value not important',
-            ],
-            'expected' => 'a73d2e7035f7fa032237c8cf0eb5be22',
-        ];
-        yield 'Configuration with both paramExpanded and URLS set' => [
-            'configuration' => [
-                'testKey' => 'testValue',
-                'paramExpanded' => 'Value not important',
-                'URLs' => 'Value not important',
-            ],
-            'expected' => 'a73d2e7035f7fa032237c8cf0eb5be22',
-        ];
-        yield 'Configuration with both paramExpanded and URLS set, will return same hash' => [
-            'configuration' => [
-                'testKey' => 'testValue',
-                'paramExpanded' => 'Value not important, but different than test case before',
-                'URLs' => 'Value not important, but different than test case before',
-            ],
-            'expected' => 'a73d2e7035f7fa032237c8cf0eb5be22',
         ];
     }
