+++ Tests/Unit/Utility/CrawlerConfigurationUtilityTest.php
@@ -1,0 +1,106 @@
+<?php
+
+declare(strict_types=1);
+
+namespace AOE\Crawler\Tests\Unit\Utility;
+
+/*
+ * (c) 2024-     Tomas Norre Mikkelsen <tomasnorre@gmail.com>
+ *
+ * This file is part of the TYPO3 Crawler Extension.
+ *
+ * It is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU General Public License, either version 2
+ * of the License, or any later version.
+ *
+ * For the full copyright and license information, please read the
+ * LICENSE.txt file that was distributed with this source code.
+ *
+ * The TYPO3 project - inspiring people to share!
+ */
+
+use AOE\Crawler\Utility\CrawlerConfigurationUtility;
+use PHPUnit\Framework\Attributes\CoversClass;
+use PHPUnit\Framework\Attributes\DataProvider;
+use PHPUnit\Framework\Attributes\Test;
+use TYPO3\TestingFramework\Core\Unit\UnitTestCase;
+
+#[CoversClass(CrawlerConfigurationUtility::class)]
+class CrawlerConfigurationUtilityTest extends UnitTestCase
+{
+    #[Test]
+    public function getConfigurationHash(): void
+    {
+        $configuration = [
+            'paramExpanded' => 'extendedParameter',
+            'URLs' => 'URLs',
+            'NotImportantParameter' => 'value not important',
+        ];
+
+        $originalCheckSum = md5(serialize($configuration));
+
+        self::assertNotNull(
+            CrawlerConfigurationUtility::getConfigurationHash($configuration)
+        );
+
+        self::assertNotEquals(
+            $originalCheckSum,
+            CrawlerConfigurationUtility::getConfigurationHash($configuration)
+        );
+
+        unset($configuration['paramExpanded'], $configuration['URLs']);
+        $newCheckSum = md5(serialize($configuration));
+        self::assertEquals($newCheckSum, CrawlerConfigurationUtility::getConfigurationHash($configuration));
+    }
+
+    #[DataProvider('getConfigurationHasReturnsExpectedValueDataProvider')]
+    #[Test]
+    public function getConfigurationHasReturnsExpectedValue(array $configuration, string $expected): void
+    {
+        self::assertEquals($expected, CrawlerConfigurationUtility::getConfigurationHash($configuration));
+    }
+
+    public static function getConfigurationHasReturnsExpectedValueDataProvider(): iterable
+    {
+        yield 'Configuration with either paramExpanded nor URLs set' => [
+            'configuration' => [
+                'testKey' => 'testValue',
+                'paramExpanded' => '',
+                'URLs' => '',
+            ],
+            'expected' => 'a73d2e7035f7fa032237c8cf0eb5be22',
+        ];
+        yield 'Configuration with only paramExpanded set' => [
+            'configuration' => [
+                'testKey' => 'testValue',
+                'paramExpanded' => 'Value not important',
+                'URLs' => '',
+            ],
+            'expected' => 'a73d2e7035f7fa032237c8cf0eb5be22',
+        ];
+        yield 'Configuration with only URLS set' => [
+            'configuration' => [
+                'testKey' => 'testValue',
+                'paramExpanded' => '',
+                'URLs' => 'Value not important',
+            ],
+            'expected' => 'a73d2e7035f7fa032237c8cf0eb5be22',
+        ];
+        yield 'Configuration with both paramExpanded and URLS set' => [
+            'configuration' => [
+                'testKey' => 'testValue',
+                'paramExpanded' => 'Value not important',
+                'URLs' => 'Value not important',
+            ],
+            'expected' => 'a73d2e7035f7fa032237c8cf0eb5be22',
+        ];
+        yield 'Configuration with both paramExpanded and URLS set, will return same hash' => [
+            'configuration' => [
+                'testKey' => 'testValue',
+                'paramExpanded' => 'Value not important, but different than test case before',
+                'URLs' => 'Value not important, but different than test case before',
+            ],
+            'expected' => 'a73d2e7035f7fa032237c8cf0eb5be22',
+        ];
+    }
+}
